requirements.txt
torch
torchvision
pyYAML

wandb

